vacuum uses sys_tdes

+++ src/transaction/log_writer.c
@@ -1729,7 +1729,7 @@ logwr_check_page_checksum (THREAD_ENTRY * thread_p, LOG_PAGE * log_pgptr)
       p += sample_nbytes;
     }
 
-  error_code = crypt_crc32 (thread_p, (char *) buf, sizeof_buf, &checksum_crc32);
+  error_code = crypt_crc32 (thread_p, (char *) buf, (int) sizeof_buf, &checksum_crc32);
 
   /* Restores the saved checksum */
   log_pgptr->hdr.checksum = saved_checksum_crc32;
